issue #70: added github-metrics to console_script
