+++ setup.py
@@ -8,6 +8,7 @@ setup(
         'console_scripts': [
             'remove-previous-images=remove_previous_image.remove_previous_image:main',
             'webtop-template=webtop_template.webtop_template:main',
+            'github-metrics=github_metrics.github_metrics:main',
         ],
     },
     url='https://github.com/ai-cfia/devops.git',
